Split player attack moves into 4 separate files
