+++ src/Main.java
@@ -12,7 +12,7 @@ import java.util.Objects;
 import java.util.Random;
 import java.util.concurrent.TimeUnit;
 
-public class Main extends com_moves
+public class Main extends play_part_4
 {
 
     MouseListener ml = new MouseListener()
@@ -20,7 +20,7 @@ public class Main extends com_moves
         @Override
         public void mouseClicked(MouseEvent e)
         {
-            if(!Objects.equals(cur_aircraft, "finished"))
+            if (!Objects.equals(cur_aircraft, "finished"))
             {
                 if (c == j1 && j1.isEnabled())
                 {
@@ -329,405 +329,306 @@ public class Main extends com_moves
                     //JOptionPane.showMessageDialog(j, "That is an invalid move - Please try another move");
                 }
             }
-            if(turn_flag == 0)
+            if (turn_flag == 0)
             {
                 if (com == l1 && l1.isEnabled())
                 {
                     l1Call();
-                }
-                else if (com == l11 && l11.isEnabled())
+                } else if (com == l11 && l11.isEnabled())
                 {
                     l11Call();
-                }
-                else if (com == l21 && l21.isEnabled())
+                } else if (com == l21 && l21.isEnabled())
                 {
                     l21Call();
-                }
-                else if (com == l31 && l31.isEnabled())
+                } else if (com == l31 && l31.isEnabled())
                 {
                     l31Call();
-                }
-                else if (com == l41 && l41.isEnabled())
+                } else if (com == l41 && l41.isEnabled())
                 {
                     l41Call();
-                }
-                else if (com == l51 && l51.isEnabled())
+                } else if (com == l51 && l51.isEnabled())
                 {
                     l51Call();
-                }
-                else if (com == l61 && l61.isEnabled())
+                } else if (com == l61 && l61.isEnabled())
                 {
                     l61Call();
-                }
-                else if (com == l71 && l71.isEnabled())
+                } else if (com == l71 && l71.isEnabled())
                 {
                     l71Call();
-                }
-                else if (com == l81 && l81.isEnabled())
+                } else if (com == l81 && l81.isEnabled())
                 {
                     l81Call();
-                }
-                else if (com == l91 && l91.isEnabled())
+                } else if (com == l91 && l91.isEnabled())
                 {
                     l91Call();
-                }
-                else if (com == l2 && l2.isEnabled())
+                } else if (com == l2 && l2.isEnabled())
                 {
                     l2Call();
-                }
-                else if (com == l12 && l12.isEnabled())
+                } else if (com == l12 && l12.isEnabled())
                 {
                     l12Call();
-                }
-                else if (com == l22 && l22.isEnabled())
+                } else if (com == l22 && l22.isEnabled())
                 {
                     l22Call();
-                }
-                else if (com == l32 && l32.isEnabled())
+                } else if (com == l32 && l32.isEnabled())
                 {
                     l32Call();
-                }
-                else if (com == l42 && l42.isEnabled())
+                } else if (com == l42 && l42.isEnabled())
                 {
                     l42Call();
-                }
-                else if (com == l52 && l52.isEnabled())
+                } else if (com == l52 && l52.isEnabled())
                 {
                     l52Call();
-                }
-                else if (com == l62 && l62.isEnabled())
+                } else if (com == l62 && l62.isEnabled())
                 {
                     l62Call();
-                }
-                else if (com == l72 && l72.isEnabled())
+                } else if (com == l72 && l72.isEnabled())
                 {
                     l72Call();
-                }
-                else if (com == l82 && l82.isEnabled())
+                } else if (com == l82 && l82.isEnabled())
                 {
                     l82Call();
-                }
-                else if (com == l92 && l92.isEnabled())
+                } else if (com == l92 && l92.isEnabled())
                 {
                     l92Call();
-                }
-                else if (com == l3 && l3.isEnabled())
+                } else if (com == l3 && l3.isEnabled())
                 {
                     l3Call();
-                }
-                else if (com == l13 && l13.isEnabled())
+                } else if (com == l13 && l13.isEnabled())
                 {
                     l13Call();
-                }
-                else if (com == l23 && l23.isEnabled())
+                } else if (com == l23 && l23.isEnabled())
                 {
                     l23Call();
-                }
-                else if (com == l33 && l33.isEnabled())
+                } else if (com == l33 && l33.isEnabled())
                 {
                     l33Call();
-                }
-                else if (com == l43 && l43.isEnabled())
+                } else if (com == l43 && l43.isEnabled())
                 {
                     l43Call();
-                }
-                else if (com == l53 && l53.isEnabled())
+                } else if (com == l53 && l53.isEnabled())
                 {
                     l53Call();
-                }
-                else if (com == l63 && l63.isEnabled())
+                } else if (com == l63 && l63.isEnabled())
                 {
                     l63Call();
-                }
-                else if (com == l73 && l73.isEnabled())
+                } else if (com == l73 && l73.isEnabled())
                 {
                     l73Call();
-                }
-                else if (com == l83 && l83.isEnabled())
+                } else if (com == l83 && l83.isEnabled())
                 {
                     l83Call();
-                }
-                else if (com == l93 && l93.isEnabled())
+                } else if (com == l93 && l93.isEnabled())
                 {
                     l93Call();
-                }
-                else if (com == l4 && l4.isEnabled())
+                } else if (com == l4 && l4.isEnabled())
                 {
                     l4Call();
-                }
-                else if (com == l14 && l14.isEnabled())
+                } else if (com == l14 && l14.isEnabled())
                 {
                     l14Call();
-                }
-                else if (com == l24 && l24.isEnabled())
+                } else if (com == l24 && l24.isEnabled())
                 {
                     l24Call();
-                }
-                else if (com == l34 && l34.isEnabled())
+                } else if (com == l34 && l34.isEnabled())
                 {
                     l34Call();
-                }
-                else if (com == l44 && l44.isEnabled())
+                } else if (com == l44 && l44.isEnabled())
                 {
                     l44Call();
-                }
-                else if (com == l54 && l54.isEnabled())
+                } else if (com == l54 && l54.isEnabled())
                 {
                     l54Call();
-                }
-                else if (com == l64 && l64.isEnabled())
+                } else if (com == l64 && l64.isEnabled())
                 {
                     l64Call();
-                }
-                else if (com == l74 && l74.isEnabled())
+                } else if (com == l74 && l74.isEnabled())
                 {
                     l74Call();
-                }
-                else if (com == l84 && l84.isEnabled())
+                } else if (com == l84 && l84.isEnabled())
                 {
                     l84Call();
-                }
-                else if (com == l94 && l94.isEnabled())
+                } else if (com == l94 && l94.isEnabled())
                 {
                     l94Call();
-                }
-                else if (com == l5 && l5.isEnabled())
+                } else if (com == l5 && l5.isEnabled())
                 {
                     l5Call();
-                }
-                else if (com == l15 && l15.isEnabled())
+                } else if (com == l15 && l15.isEnabled())
                 {
                     l15Call();
-                }
-                else if (com == l25 && l25.isEnabled())
+                } else if (com == l25 && l25.isEnabled())
                 {
                     l25Call();
-                }
-                else if (com == l35 && l35.isEnabled())
+                } else if (com == l35 && l35.isEnabled())
                 {
                     l35Call();
-                }
-                else if (com == l45 && l45.isEnabled())
+                } else if (com == l45 && l45.isEnabled())
                 {
                     l45Call();
-                }
-                else if (com == l55 && l55.isEnabled())
+                } else if (com == l55 && l55.isEnabled())
                 {
                     l55Call();
-                }
-                else if (com == l65 && l65.isEnabled())
+                } else if (com == l65 && l65.isEnabled())
                 {
                     l65Call();
-                }
-                else if (com == l75 && l75.isEnabled())
+                } else if (com == l75 && l75.isEnabled())
                 {
                     l75Call();
-                }
-                else if (com == l85 && l85.isEnabled())
+                } else if (com == l85 && l85.isEnabled())
                 {
                     l85Call();
-                }
-                else if (com == l95 && l95.isEnabled())
+                } else if (com == l95 && l95.isEnabled())
                 {
                     l95Call();
-                }
-                else if (com == l6 && l6.isEnabled())
+                } else if (com == l6 && l6.isEnabled())
                 {
                     l6Call();
-                }
-                else if (com == l16 && l16.isEnabled())
+                } else if (com == l16 && l16.isEnabled())
                 {
                     l16Call();
-                }
-                else if (com == l26 && l26.isEnabled())
+                } else if (com == l26 && l26.isEnabled())
                 {
                     l26Call();
-                }
-                else if (com == l36 && l36.isEnabled())
+                } else if (com == l36 && l36.isEnabled())
                 {
                     l36Call();
-                }
-                else if (com == l46 && l46.isEnabled())
+                } else if (com == l46 && l46.isEnabled())
                 {
                     l46Call();
-                }
-                else if (com == l56 && l56.isEnabled())
+                } else if (com == l56 && l56.isEnabled())
                 {
                     l56Call();
-                }
-                else if (com == l66 && l66.isEnabled())
+                } else if (com == l66 && l66.isEnabled())
                 {
                     l66Call();
-                }
-                else if (com == l76 && l76.isEnabled())
+                } else if (com == l76 && l76.isEnabled())
                 {
                     l76Call();
-                }
-                else if (com == l86 && l86.isEnabled())
+                } else if (com == l86 && l86.isEnabled())
                 {
                     l86Call();
-                }
-                else if (com == l96 && l96.isEnabled())
+                } else if (com == l96 && l96.isEnabled())
                 {
                     l96Call();
-                }
-                else if (com == l7 && l7.isEnabled())
+                } else if (com == l7 && l7.isEnabled())
                 {
                     l7Call();
-                }
-                else if (com == l17 && l17.isEnabled())
+                } else if (com == l17 && l17.isEnabled())
                 {
                     l17Call();
-                }
-                else if (com == l27 && l27.isEnabled())
+                } else if (com == l27 && l27.isEnabled())
                 {
                     l27Call();
-                }
-                else if (com == l37 && l37.isEnabled())
+                } else if (com == l37 && l37.isEnabled())
                 {
                     l37Call();
-                }
-                else if (com == l47 && l47.isEnabled())
+                } else if (com == l47 && l47.isEnabled())
                 {
                     l47Call();
-                }
-                else if (com == l57 && l57.isEnabled())
+                } else if (com == l57 && l57.isEnabled())
                 {
                     l57Call();
-                }
-                else if (com == l67 && l67.isEnabled())
+                } else if (com == l67 && l67.isEnabled())
                 {
                     l67Call();
-                }
-                else if (com == l77 && l77.isEnabled())
+                } else if (com == l77 && l77.isEnabled())
                 {
                     l77Call();
-                }
-                else if (com == l87 && l87.isEnabled())
+                } else if (com == l87 && l87.isEnabled())
                 {
                     l87Call();
-                }
-                else if (com == l97 && l97.isEnabled())
+                } else if (com == l97 && l97.isEnabled())
                 {
                     l97Call();
-                }
-                else if (com == l8 && l8.isEnabled())
+                } else if (com == l8 && l8.isEnabled())
                 {
                     l8Call();
-                }
-                else if (com == l18 && l18.isEnabled())
+                } else if (com == l18 && l18.isEnabled())
                 {
                     l18Call();
-                }
-                else if (com == l28 && l28.isEnabled())
+                } else if (com == l28 && l28.isEnabled())
                 {
                     l28Call();
-                }
-                else if (com == l38 && l38.isEnabled())
+                } else if (com == l38 && l38.isEnabled())
                 {
                     l38Call();
-                }
-                else if (com == l48 && l48.isEnabled())
+                } else if (com == l48 && l48.isEnabled())
                 {
                     l48Call();
-                }
-                else if (com == l58 && l58.isEnabled())
+                } else if (com == l58 && l58.isEnabled())
                 {
                     l58Call();
-                }
-                else if (com == l68 && l68.isEnabled())
+                } else if (com == l68 && l68.isEnabled())
                 {
                     l68Call();
-                }
-                else if (com == l78 && l78.isEnabled())
+                } else if (com == l78 && l78.isEnabled())
                 {
                     l78Call();
-                }
-                else if (com == l88 && l88.isEnabled())
+                } else if (com == l88 && l88.isEnabled())
                 {
                     l88Call();
-                }
-                else if (com == l98 && l98.isEnabled())
+                } else if (com == l98 && l98.isEnabled())
                 {
                     l98Call();
-                }
-                else if (com == l9 && l9.isEnabled())
+                } else if (com == l9 && l9.isEnabled())
                 {
                     l9Call();
-                }
-                else if (com == l19 && l19.isEnabled())
+                } else if (com == l19 && l19.isEnabled())
                 {
                     l19Call();
-                }
-                else if (com == l29 && l29.isEnabled())
+                } else if (com == l29 && l29.isEnabled())
                 {
                     l29Call();
-                }
-                else if (com == l39 && l39.isEnabled())
+                } else if (com == l39 && l39.isEnabled())
                 {
                     l39Call();
-                }
-                else if (com == l49 && l49.isEnabled())
+                } else if (com == l49 && l49.isEnabled())
                 {
                     l49Call();
-                }
-                else if (com == l59 && l59.isEnabled())
+                } else if (com == l59 && l59.isEnabled())
                 {
                     l59Call();
-                }
-                else if (com == l69 && l69.isEnabled())
+                } else if (com == l69 && l69.isEnabled())
                 {
                     l69Call();
-                }
-                else if (com == l79 && l79.isEnabled())
+                } else if (com == l79 && l79.isEnabled())
                 {
                     l79Call();
-                }
-                else if (com == l89 && l89.isEnabled())
+                } else if (com == l89 && l89.isEnabled())
                 {
                     l89Call();
-                }
-                else if (com == l99 && l99.isEnabled())
+                } else if (com == l99 && l99.isEnabled())
                 {
                     l99Call();
-                }
-                else if (com == l10 && l10.isEnabled())
+                } else if (com == l10 && l10.isEnabled())
                 {
                     l10Call();
-                }
-                else if (com == l20 && l20.isEnabled())
+                } else if (com == l20 && l20.isEnabled())
                 {
                     l20Call();
-                }
-                else if (com == l30 && l30.isEnabled())
+                } else if (com == l30 && l30.isEnabled())
                 {
                     l30Call();
-                }
-                else if (com == l40 && l40.isEnabled())
+                } else if (com == l40 && l40.isEnabled())
                 {
                     l40Call();
-                }
-                else if (com == l50 && l50.isEnabled())
+                } else if (com == l50 && l50.isEnabled())
                 {
                     l50Call();
-                }
-                else if (com == l60 && l60.isEnabled())
+                } else if (com == l60 && l60.isEnabled())
                 {
                     l60Call();
-                }
-                else if (com == l70 && l70.isEnabled())
+                } else if (com == l70 && l70.isEnabled())
                 {
                     l70Call();
-                }
-                else if (com == l80 && l80.isEnabled())
+                } else if (com == l80 && l80.isEnabled())
                 {
                     l80Call();
-                }
-                else if (com == l90 && l90.isEnabled())
+                } else if (com == l90 && l90.isEnabled())
                 {
                     l90Call();
-                }
-                else if (com == l100 && l100.isEnabled())
+                } else if (com == l100 && l100.isEnabled())
                 {
                     l100Call();
                 }
@@ -758,6225 +659,6 @@ public class Main extends com_moves
 
         }
     };
-
-
-
-
-
-    public void l1Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 1)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 1)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 1)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 1)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 1)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l1.setIcon(new ImageIcon("water_ship_hit.png"));
-            l1.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l1.setIcon(new ImageIcon("water_ship_miss.png"));
-            l1.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l1.setEnabled(false);
-        who_is_playing();
-    }
-
-    public void l2Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 2)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 2)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 2)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 2)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 2)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l2.setIcon(new ImageIcon("water_ship_hit.png"));
-            l2.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l2.setIcon(new ImageIcon("water_ship_miss.png"));
-            l2.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l2.setEnabled(false);
-        who_is_playing();
-    }
-    public void l3Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 3)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 3)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 3)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 3)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 3)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l3.setIcon(new ImageIcon("water_ship_hit.png"));
-            l3.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l3.setIcon(new ImageIcon("water_ship_miss.png"));
-            l3.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l3.setEnabled(false);
-        who_is_playing();
-    }
-    public void l4Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 4)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 4)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 4)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 4)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 4)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l4.setIcon(new ImageIcon("water_ship_hit.png"));
-            l4.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l4.setIcon(new ImageIcon("water_ship_miss.png"));
-            l4.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l4.setEnabled(false);
-        who_is_playing();
-    }
-    public void l5Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 5)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 5)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 5)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 5)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 5)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l5.setIcon(new ImageIcon("water_ship_hit.png"));
-            l5.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l5.setIcon(new ImageIcon("water_ship_miss.png"));
-            l5.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l5.setEnabled(false);
-        who_is_playing();
-    }
-    public void l6Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 6)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 6)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 6)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 6)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 6)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l6.setIcon(new ImageIcon("water_ship_hit.png"));
-            l6.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l6.setIcon(new ImageIcon("water_ship_miss.png"));
-            l6.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l6.setEnabled(false);
-        who_is_playing();
-    }
-    public void l7Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 7)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 7)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 7)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 7)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 7)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l7.setIcon(new ImageIcon("water_ship_hit.png"));
-            l7.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l7.setIcon(new ImageIcon("water_ship_miss.png"));
-            l7.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l7.setEnabled(false);
-        who_is_playing();
-    }
-    public void l8Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 8)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 8)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 8)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 8)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 8)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l8.setIcon(new ImageIcon("water_ship_hit.png"));
-            l8.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l8.setIcon(new ImageIcon("water_ship_miss.png"));
-            l8.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l8.setEnabled(false);
-        who_is_playing();
-    }
-    public void l9Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 9)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 9)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 9)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 9)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 9)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l9.setIcon(new ImageIcon("water_ship_hit.png"));
-            l9.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l9.setIcon(new ImageIcon("water_ship_miss.png"));
-            l9.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l9.setEnabled(false);
-        who_is_playing();
-    }
-    public void l10Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 10)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 10)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 10)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 10)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 10)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l10.setIcon(new ImageIcon("water_ship_hit.png"));
-            l10.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l10.setIcon(new ImageIcon("water_ship_miss.png"));
-            l10.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l10.setEnabled(false);
-        who_is_playing();
-    }
-    public void l11Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 11)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 11)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 11)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 11)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 11)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l11.setIcon(new ImageIcon("water_ship_hit.png"));
-            l11.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l11.setIcon(new ImageIcon("water_ship_miss.png"));
-            l11.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l11.setEnabled(false);
-        who_is_playing();
-    }
-
-    public void l12Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 12)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 12)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 12)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 12)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 12)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l12.setIcon(new ImageIcon("water_ship_hit.png"));
-            l12.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l12.setIcon(new ImageIcon("water_ship_miss.png"));
-            l12.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l12.setEnabled(false);
-        who_is_playing();
-    }
-    public void l13Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 13)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 13)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 13)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 13)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 13)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l13.setIcon(new ImageIcon("water_ship_hit.png"));
-            l13.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l13.setIcon(new ImageIcon("water_ship_miss.png"));
-            l13.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l13.setEnabled(false);
-        who_is_playing();
-    }
-    public void l14Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 14)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 14)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 14)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 14)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 14)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l14.setIcon(new ImageIcon("water_ship_hit.png"));
-            l14.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l14.setIcon(new ImageIcon("water_ship_miss.png"));
-            l14.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l14.setEnabled(false);
-        who_is_playing();
-    }
-    public void l15Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 15)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 15)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 15)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 15)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 15)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l15.setIcon(new ImageIcon("water_ship_hit.png"));
-            l15.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l15.setIcon(new ImageIcon("water_ship_miss.png"));
-            l15.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l15.setEnabled(false);
-        who_is_playing();
-    }
-    public void l16Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 16)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 16)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 16)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 16)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 16)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l16.setIcon(new ImageIcon("water_ship_hit.png"));
-            l16.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l16.setIcon(new ImageIcon("water_ship_miss.png"));
-            l16.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l16.setEnabled(false);
-        who_is_playing();
-    }
-    public void l17Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 17)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 17)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 17)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 17)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 17)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l17.setIcon(new ImageIcon("water_ship_hit.png"));
-            l17.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l17.setIcon(new ImageIcon("water_ship_miss.png"));
-            l17.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l17.setEnabled(false);
-        who_is_playing();
-    }
-    public void l18Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 18)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 18)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 18)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 18)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 18)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l18.setIcon(new ImageIcon("water_ship_hit.png"));
-            l18.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l18.setIcon(new ImageIcon("water_ship_miss.png"));
-            l18.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l18.setEnabled(false);
-        who_is_playing();
-    }
-    public void l19Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 19)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 19)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 19)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 19)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 19)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l19.setIcon(new ImageIcon("water_ship_hit.png"));
-            l19.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l19.setIcon(new ImageIcon("water_ship_miss.png"));
-            l19.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l19.setEnabled(false);
-        who_is_playing();
-    }
-    public void l20Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 20)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 20)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 20)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 20)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 20)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l20.setIcon(new ImageIcon("water_ship_hit.png"));
-            l20.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l20.setIcon(new ImageIcon("water_ship_miss.png"));
-            l20.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l20.setEnabled(false);
-        who_is_playing();
-    }
-    public void l21Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 21)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 21)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 21)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 21)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 21)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l21.setIcon(new ImageIcon("water_ship_hit.png"));
-            l21.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l21.setIcon(new ImageIcon("water_ship_miss.png"));
-            l21.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l21.setEnabled(false);
-        who_is_playing();
-    }
-
-    public void l22Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 22)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 22)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 22)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 22)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 22)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l22.setIcon(new ImageIcon("water_ship_hit.png"));
-            l22.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l22.setIcon(new ImageIcon("water_ship_miss.png"));
-            l22.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l22.setEnabled(false);
-        who_is_playing();
-    }
-    public void l23Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 23)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 23)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 23)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 23)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 23)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l23.setIcon(new ImageIcon("water_ship_hit.png"));
-            l23.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l23.setIcon(new ImageIcon("water_ship_miss.png"));
-            l23.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l23.setEnabled(false);
-        who_is_playing();
-    }
-    public void l24Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 24)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 24)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 24)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 24)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 24)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l24.setIcon(new ImageIcon("water_ship_hit.png"));
-            l24.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l24.setIcon(new ImageIcon("water_ship_miss.png"));
-            l24.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l24.setEnabled(false);
-        who_is_playing();
-    }
-    public void l25Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 25)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 25)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 25)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 25)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 25)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l25.setIcon(new ImageIcon("water_ship_hit.png"));
-            l25.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l25.setIcon(new ImageIcon("water_ship_miss.png"));
-            l25.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l25.setEnabled(false);
-        who_is_playing();
-    }
-    public void l26Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 26)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 26)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 26)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 26)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 26)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l26.setIcon(new ImageIcon("water_ship_hit.png"));
-            l26.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l26.setIcon(new ImageIcon("water_ship_miss.png"));
-            l26.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l26.setEnabled(false);
-        who_is_playing();
-    }
-    public void l27Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 27)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 27)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 27)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 27)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 27)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l27.setIcon(new ImageIcon("water_ship_hit.png"));
-            l27.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l27.setIcon(new ImageIcon("water_ship_miss.png"));
-            l27.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l27.setEnabled(false);
-        who_is_playing();
-    }
-    public void l28Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 28)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 28)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 28)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 28)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 28)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l28.setIcon(new ImageIcon("water_ship_hit.png"));
-            l28.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l28.setIcon(new ImageIcon("water_ship_miss.png"));
-            l28.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l28.setEnabled(false);
-        who_is_playing();
-    }
-    public void l29Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 29)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 29)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 29)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 29)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 29)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l29.setIcon(new ImageIcon("water_ship_hit.png"));
-            l29.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l29.setIcon(new ImageIcon("water_ship_miss.png"));
-            l29.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l29.setEnabled(false);
-        who_is_playing();
-    }
-    public void l30Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 30)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 30)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 30)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 30)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 30)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l30.setIcon(new ImageIcon("water_ship_hit.png"));
-            l30.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l30.setIcon(new ImageIcon("water_ship_miss.png"));
-            l30.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l30.setEnabled(false);
-        who_is_playing();
-    }
-    public void l31Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 31)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 31)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 31)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 31)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 31)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l31.setIcon(new ImageIcon("water_ship_hit.png"));
-            l31.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l31.setIcon(new ImageIcon("water_ship_miss.png"));
-            l31.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l31.setEnabled(false);
-        who_is_playing();
-    }
-
-    public void l32Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 32)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 32)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 32)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 32)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 32)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l32.setIcon(new ImageIcon("water_ship_hit.png"));
-            l32.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l32.setIcon(new ImageIcon("water_ship_miss.png"));
-            l32.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l32.setEnabled(false);
-        who_is_playing();
-    }
-    public void l33Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 33)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 33)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 33)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 33)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 33)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l33.setIcon(new ImageIcon("water_ship_hit.png"));
-            l33.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l33.setIcon(new ImageIcon("water_ship_miss.png"));
-            l33.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l33.setEnabled(false);
-        who_is_playing();
-    }
-    public void l34Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 34)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 34)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 34)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 34)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 34)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l34.setIcon(new ImageIcon("water_ship_hit.png"));
-            l34.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l34.setIcon(new ImageIcon("water_ship_miss.png"));
-            l34.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l34.setEnabled(false);
-        who_is_playing();
-    }
-    public void l35Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 35)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 35)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 35)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 35)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 35)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l35.setIcon(new ImageIcon("water_ship_hit.png"));
-            l35.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l35.setIcon(new ImageIcon("water_ship_miss.png"));
-            l35.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l35.setEnabled(false);
-        who_is_playing();
-    }
-    public void l36Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 36)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 36)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 36)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 36)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 36)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l36.setIcon(new ImageIcon("water_ship_hit.png"));
-            l36.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l36.setIcon(new ImageIcon("water_ship_miss.png"));
-            l36.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l36.setEnabled(false);
-        who_is_playing();
-    }
-    public void l37Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 37)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 37)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 37)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 37)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 37)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l37.setIcon(new ImageIcon("water_ship_hit.png"));
-            l37.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l37.setIcon(new ImageIcon("water_ship_miss.png"));
-            l37.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l37.setEnabled(false);
-        who_is_playing();
-    }
-    public void l38Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 38)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 38)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 38)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 38)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 38)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l38.setIcon(new ImageIcon("water_ship_hit.png"));
-            l38.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l38.setIcon(new ImageIcon("water_ship_miss.png"));
-            l38.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l38.setEnabled(false);
-        who_is_playing();
-    }
-    public void l39Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 39)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 39)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 39)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 39)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 39)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l39.setIcon(new ImageIcon("water_ship_hit.png"));
-            l39.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l39.setIcon(new ImageIcon("water_ship_miss.png"));
-            l39.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l39.setEnabled(false);
-        who_is_playing();
-    }
-    public void l40Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 40)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 40)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 40)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 40)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 40)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l40.setIcon(new ImageIcon("water_ship_hit.png"));
-            l40.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l40.setIcon(new ImageIcon("water_ship_miss.png"));
-            l40.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l40.setEnabled(false);
-        who_is_playing();
-    }
-    public void l41Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 41)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 41)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 41)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 41)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 41)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l41.setIcon(new ImageIcon("water_ship_hit.png"));
-            l41.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l41.setIcon(new ImageIcon("water_ship_miss.png"));
-            l41.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l41.setEnabled(false);
-        who_is_playing();
-    }
-
-    public void l42Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 42)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 42)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 42)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 42)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 42)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l42.setIcon(new ImageIcon("water_ship_hit.png"));
-            l42.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l42.setIcon(new ImageIcon("water_ship_miss.png"));
-            l42.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l42.setEnabled(false);
-        who_is_playing();
-    }
-    public void l43Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 43)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 43)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 43)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 43)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 43)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l43.setIcon(new ImageIcon("water_ship_hit.png"));
-            l43.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l43.setIcon(new ImageIcon("water_ship_miss.png"));
-            l43.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l43.setEnabled(false);
-        who_is_playing();
-    }
-    public void l44Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 44)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 44)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 44)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 44)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 44)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l44.setIcon(new ImageIcon("water_ship_hit.png"));
-            l44.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l44.setIcon(new ImageIcon("water_ship_miss.png"));
-            l44.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l44.setEnabled(false);
-        who_is_playing();
-    }
-    public void l45Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 45)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 45)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 45)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 45)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 45)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l45.setIcon(new ImageIcon("water_ship_hit.png"));
-            l45.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l45.setIcon(new ImageIcon("water_ship_miss.png"));
-            l45.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l45.setEnabled(false);
-        who_is_playing();
-    }
-    public void l46Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 46)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 46)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 46)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 46)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 46)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l46.setIcon(new ImageIcon("water_ship_hit.png"));
-            l46.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l46.setIcon(new ImageIcon("water_ship_miss.png"));
-            l46.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l46.setEnabled(false);
-        who_is_playing();
-    }
-    public void l47Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 47)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 47)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 47)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 47)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 47)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l47.setIcon(new ImageIcon("water_ship_hit.png"));
-            l47.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l47.setIcon(new ImageIcon("water_ship_miss.png"));
-            l47.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l47.setEnabled(false);
-        who_is_playing();
-    }
-    public void l48Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 48)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 48)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 48)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 48)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 48)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l48.setIcon(new ImageIcon("water_ship_hit.png"));
-            l48.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l48.setIcon(new ImageIcon("water_ship_miss.png"));
-            l48.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l48.setEnabled(false);
-        who_is_playing();
-    }
-    public void l49Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 49)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 49)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 49)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 49)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 49)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l49.setIcon(new ImageIcon("water_ship_hit.png"));
-            l49.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l49.setIcon(new ImageIcon("water_ship_miss.png"));
-            l49.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l49.setEnabled(false);
-        who_is_playing();
-    }
-    public void l50Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 50)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 50)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 50)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 50)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 50)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l50.setIcon(new ImageIcon("water_ship_hit.png"));
-            l50.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l50.setIcon(new ImageIcon("water_ship_miss.png"));
-            l50.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l50.setEnabled(false);
-        who_is_playing();
-    }
-    public void l51Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 51)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 51)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 51)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 51)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 51)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l51.setIcon(new ImageIcon("water_ship_hit.png"));
-            l51.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l51.setIcon(new ImageIcon("water_ship_miss.png"));
-            l51.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l51.setEnabled(false);
-        who_is_playing();
-    }
-
-    public void l52Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 52)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 52)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 52)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 52)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 52)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l52.setIcon(new ImageIcon("water_ship_hit.png"));
-            l52.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l52.setIcon(new ImageIcon("water_ship_miss.png"));
-            l52.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l52.setEnabled(false);
-        who_is_playing();
-    }
-    public void l53Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 53)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 53)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 53)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 53)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 53)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l53.setIcon(new ImageIcon("water_ship_hit.png"));
-            l53.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l53.setIcon(new ImageIcon("water_ship_miss.png"));
-            l53.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l53.setEnabled(false);
-        who_is_playing();
-    }
-    public void l54Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 54)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 54)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 54)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 54)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 54)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l54.setIcon(new ImageIcon("water_ship_hit.png"));
-            l54.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l54.setIcon(new ImageIcon("water_ship_miss.png"));
-            l54.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l54.setEnabled(false);
-        who_is_playing();
-    }
-    public void l55Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 55)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 55)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 55)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 55)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 55)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l55.setIcon(new ImageIcon("water_ship_hit.png"));
-            l55.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l55.setIcon(new ImageIcon("water_ship_miss.png"));
-            l55.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l55.setEnabled(false);
-        who_is_playing();
-    }
-    public void l56Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 56)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 56)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 56)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 56)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 56)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l56.setIcon(new ImageIcon("water_ship_hit.png"));
-            l56.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l56.setIcon(new ImageIcon("water_ship_miss.png"));
-            l56.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l56.setEnabled(false);
-        who_is_playing();
-    }
-    public void l57Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 57)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 57)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 57)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 57)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 57)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l57.setIcon(new ImageIcon("water_ship_hit.png"));
-            l57.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l57.setIcon(new ImageIcon("water_ship_miss.png"));
-            l57.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l57.setEnabled(false);
-        who_is_playing();
-    }
-    public void l58Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 58)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 58)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 58)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 58)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 58)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l58.setIcon(new ImageIcon("water_ship_hit.png"));
-            l58.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l58.setIcon(new ImageIcon("water_ship_miss.png"));
-            l58.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l58.setEnabled(false);
-        who_is_playing();
-    }
-    public void l59Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 59)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 59)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 59)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 59)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 59)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l59.setIcon(new ImageIcon("water_ship_hit.png"));
-            l59.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l59.setIcon(new ImageIcon("water_ship_miss.png"));
-            l59.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l59.setEnabled(false);
-        who_is_playing();
-    }
-    public void l60Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 60)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 60)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 60)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 60)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 60)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l60.setIcon(new ImageIcon("water_ship_hit.png"));
-            l60.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l60.setIcon(new ImageIcon("water_ship_miss.png"));
-            l60.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l60.setEnabled(false);
-        who_is_playing();
-    }
-    public void l61Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 61)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 61)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 61)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 61)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 61)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l61.setIcon(new ImageIcon("water_ship_hit.png"));
-            l61.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l61.setIcon(new ImageIcon("water_ship_miss.png"));
-            l61.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l61.setEnabled(false);
-        who_is_playing();
-    }
-
-    public void l62Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 62)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 62)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 62)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 62)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 62)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l62.setIcon(new ImageIcon("water_ship_hit.png"));
-            l62.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l62.setIcon(new ImageIcon("water_ship_miss.png"));
-            l62.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l62.setEnabled(false);
-        who_is_playing();
-    }
-    public void l63Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 63)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 63)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 63)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 63)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 63)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l63.setIcon(new ImageIcon("water_ship_hit.png"));
-            l63.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l63.setIcon(new ImageIcon("water_ship_miss.png"));
-            l63.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l63.setEnabled(false);
-        who_is_playing();
-    }
-    public void l64Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 64)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 64)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 64)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 64)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 64)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l64.setIcon(new ImageIcon("water_ship_hit.png"));
-            l64.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l64.setIcon(new ImageIcon("water_ship_miss.png"));
-            l64.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l64.setEnabled(false);
-        who_is_playing();
-    }
-    public void l65Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 65)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 65)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 65)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 65)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 65)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l65.setIcon(new ImageIcon("water_ship_hit.png"));
-            l65.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l65.setIcon(new ImageIcon("water_ship_miss.png"));
-            l65.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l65.setEnabled(false);
-        who_is_playing();
-    }
-    public void l66Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 66)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 66)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 66)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 66)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 66)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l66.setIcon(new ImageIcon("water_ship_hit.png"));
-            l66.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l66.setIcon(new ImageIcon("water_ship_miss.png"));
-            l66.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l66.setEnabled(false);
-        who_is_playing();
-    }
-    public void l67Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 67)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 67)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 67)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 67)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 67)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l67.setIcon(new ImageIcon("water_ship_hit.png"));
-            l67.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l67.setIcon(new ImageIcon("water_ship_miss.png"));
-            l67.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l67.setEnabled(false);
-        who_is_playing();
-    }
-    public void l68Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 68)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 68)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 68)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 68)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 68)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l68.setIcon(new ImageIcon("water_ship_hit.png"));
-            l68.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l68.setIcon(new ImageIcon("water_ship_miss.png"));
-            l68.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l68.setEnabled(false);
-        who_is_playing();
-    }
-    public void l69Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 69)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 69)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 69)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 69)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 69)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l69.setIcon(new ImageIcon("water_ship_hit.png"));
-            l69.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l69.setIcon(new ImageIcon("water_ship_miss.png"));
-            l69.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l69.setEnabled(false);
-        who_is_playing();
-    }
-    public void l70Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 70)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 70)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 70)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 70)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 70)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l70.setIcon(new ImageIcon("water_ship_hit.png"));
-            l70.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l70.setIcon(new ImageIcon("water_ship_miss.png"));
-            l70.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l70.setEnabled(false);
-        who_is_playing();
-    }
-    public void l71Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 71)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 71)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 71)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 71)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 71)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l71.setIcon(new ImageIcon("water_ship_hit.png"));
-            l71.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l71.setIcon(new ImageIcon("water_ship_miss.png"));
-            l71.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l71.setEnabled(false);
-        who_is_playing();
-    }
-
-    public void l72Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 72)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 72)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 72)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 72)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 72)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l72.setIcon(new ImageIcon("water_ship_hit.png"));
-            l72.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l72.setIcon(new ImageIcon("water_ship_miss.png"));
-            l72.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l72.setEnabled(false);
-        who_is_playing();
-    }
-    public void l73Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 73)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 73)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 73)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 73)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 73)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l73.setIcon(new ImageIcon("water_ship_hit.png"));
-            l73.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l73.setIcon(new ImageIcon("water_ship_miss.png"));
-            l73.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l73.setEnabled(false);
-        who_is_playing();
-    }
-    public void l74Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 74)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 74)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 74)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 74)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 74)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l74.setIcon(new ImageIcon("water_ship_hit.png"));
-            l74.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l74.setIcon(new ImageIcon("water_ship_miss.png"));
-            l74.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l74.setEnabled(false);
-        who_is_playing();
-    }
-    public void l75Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 75)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 75)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 75)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 75)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 75)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l75.setIcon(new ImageIcon("water_ship_hit.png"));
-            l75.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l75.setIcon(new ImageIcon("water_ship_miss.png"));
-            l75.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l75.setEnabled(false);
-        who_is_playing();
-    }
-    public void l76Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 76)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 76)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 76)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 76)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 76)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l76.setIcon(new ImageIcon("water_ship_hit.png"));
-            l76.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l76.setIcon(new ImageIcon("water_ship_miss.png"));
-            l76.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l76.setEnabled(false);
-        who_is_playing();
-    }
-    public void l77Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 77)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 77)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 77)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 77)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 77)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l77.setIcon(new ImageIcon("water_ship_hit.png"));
-            l77.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l77.setIcon(new ImageIcon("water_ship_miss.png"));
-            l77.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l77.setEnabled(false);
-        who_is_playing();
-    }
-    public void l78Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 78)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 78)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 78)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 78)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 78)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l78.setIcon(new ImageIcon("water_ship_hit.png"));
-            l78.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l78.setIcon(new ImageIcon("water_ship_miss.png"));
-            l78.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l78.setEnabled(false);
-        who_is_playing();
-    }
-    public void l79Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 79)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 79)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 79)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 79)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 79)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l79.setIcon(new ImageIcon("water_ship_hit.png"));
-            l79.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l79.setIcon(new ImageIcon("water_ship_miss.png"));
-            l79.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l79.setEnabled(false);
-        who_is_playing();
-    }
-    public void l80Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 80)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 80)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 80)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 80)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 80)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l80.setIcon(new ImageIcon("water_ship_hit.png"));
-            l80.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l80.setIcon(new ImageIcon("water_ship_miss.png"));
-            l80.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l80.setEnabled(false);
-        who_is_playing();
-    }
-    public void l81Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 81)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 81)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 81)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 81)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 81)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l81.setIcon(new ImageIcon("water_ship_hit.png"));
-            l81.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l81.setIcon(new ImageIcon("water_ship_miss.png"));
-            l81.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l81.setEnabled(false);
-        who_is_playing();
-    }
-
-    public void l82Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 82)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 82)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 82)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 82)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 82)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l82.setIcon(new ImageIcon("water_ship_hit.png"));
-            l82.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l82.setIcon(new ImageIcon("water_ship_miss.png"));
-            l82.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l82.setEnabled(false);
-        who_is_playing();
-    }
-    public void l83Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 83)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 83)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 83)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 83)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 83)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l83.setIcon(new ImageIcon("water_ship_hit.png"));
-            l83.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l83.setIcon(new ImageIcon("water_ship_miss.png"));
-            l83.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l83.setEnabled(false);
-        who_is_playing();
-    }
-    public void l84Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 84)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 84)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 84)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 84)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 84)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l84.setIcon(new ImageIcon("water_ship_hit.png"));
-            l84.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l84.setIcon(new ImageIcon("water_ship_miss.png"));
-            l84.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l84.setEnabled(false);
-        who_is_playing();
-    }
-    public void l85Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 85)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 85)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 85)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 85)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 85)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l85.setIcon(new ImageIcon("water_ship_hit.png"));
-            l85.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l85.setIcon(new ImageIcon("water_ship_miss.png"));
-            l85.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l85.setEnabled(false);
-        who_is_playing();
-    }
-    public void l86Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 86)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 86)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 86)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 86)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 86)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l86.setIcon(new ImageIcon("water_ship_hit.png"));
-            l86.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l86.setIcon(new ImageIcon("water_ship_miss.png"));
-            l86.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l86.setEnabled(false);
-        who_is_playing();
-    }
-    public void l87Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 87)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 87)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 87)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 87)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 87)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l87.setIcon(new ImageIcon("water_ship_hit.png"));
-            l87.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l87.setIcon(new ImageIcon("water_ship_miss.png"));
-            l87.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l87.setEnabled(false);
-        who_is_playing();
-    }
-    public void l88Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 88)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 88)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 88)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 88)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 88)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l88.setIcon(new ImageIcon("water_ship_hit.png"));
-            l88.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l88.setIcon(new ImageIcon("water_ship_miss.png"));
-            l88.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l88.setEnabled(false);
-        who_is_playing();
-    }
-    public void l89Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 89)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 89)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 89)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 89)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 89)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l89.setIcon(new ImageIcon("water_ship_hit.png"));
-            l89.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l89.setIcon(new ImageIcon("water_ship_miss.png"));
-            l89.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l89.setEnabled(false);
-        who_is_playing();
-    }
-    public void l90Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 90)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 90)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 90)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 90)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 90)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l90.setIcon(new ImageIcon("water_ship_hit.png"));
-            l90.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l90.setIcon(new ImageIcon("water_ship_miss.png"));
-            l90.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l90.setEnabled(false);
-        who_is_playing();
-    }
-    public void l91Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 91)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 91)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 91)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 91)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 91)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l91.setIcon(new ImageIcon("water_ship_hit.png"));
-            l91.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l91.setIcon(new ImageIcon("water_ship_miss.png"));
-            l91.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l91.setEnabled(false);
-        who_is_playing();
-    }
-
-    public void l92Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 92)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 92)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 92)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 92)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 92)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l92.setIcon(new ImageIcon("water_ship_hit.png"));
-            l92.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l92.setIcon(new ImageIcon("water_ship_miss.png"));
-            l92.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l92.setEnabled(false);
-        who_is_playing();
-    }
-    public void l93Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 93)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 93)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 93)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 93)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 93)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l93.setIcon(new ImageIcon("water_ship_hit.png"));
-            l93.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l93.setIcon(new ImageIcon("water_ship_miss.png"));
-            l93.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l93.setEnabled(false);
-        who_is_playing();
-    }
-    public void l94Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 94)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 94)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 94)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 94)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 94)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l94.setIcon(new ImageIcon("water_ship_hit.png"));
-            l94.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l94.setIcon(new ImageIcon("water_ship_miss.png"));
-            l94.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l94.setEnabled(false);
-        who_is_playing();
-    }
-    public void l95Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 95)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 95)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 95)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 95)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 95)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l95.setIcon(new ImageIcon("water_ship_hit.png"));
-            l95.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l95.setIcon(new ImageIcon("water_ship_miss.png"));
-            l95.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l95.setEnabled(false);
-        who_is_playing();
-    }
-    public void l96Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 96)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 96)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 96)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 96)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 96)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l96.setIcon(new ImageIcon("water_ship_hit.png"));
-            l96.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l96.setIcon(new ImageIcon("water_ship_miss.png"));
-            l96.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l96.setEnabled(false);
-        who_is_playing();
-    }
-    public void l97Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 97)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 97)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 97)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 97)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 97)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l97.setIcon(new ImageIcon("water_ship_hit.png"));
-            l97.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l97.setIcon(new ImageIcon("water_ship_miss.png"));
-            l97.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l97.setEnabled(false);
-        who_is_playing();
-    }
-    public void l98Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 98)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 98)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 98)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 98)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 98)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l98.setIcon(new ImageIcon("water_ship_hit.png"));
-            l98.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l98.setIcon(new ImageIcon("water_ship_miss.png"));
-            l98.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l98.setEnabled(false);
-        who_is_playing();
-    }
-    public void l99Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 99)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 99)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 99)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 99)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 99)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l99.setIcon(new ImageIcon("water_ship_hit.png"));
-            l99.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l99.setIcon(new ImageIcon("water_ship_miss.png"));
-            l99.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l99.setEnabled(false);
-        who_is_playing();
-    }
-    public void l100Call()
-    {
-        int flag = 0;
-        for(int i = 0 ; i < 5 ; i ++)
-        {
-            if(com_aircraft_pos[i] == 100)
-            {
-                play_ai_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 4 ; i ++)
-        {
-            if(com_battleship_pos[i] == 100)
-            {
-                play_bat_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_submarine_pos[i] == 100)
-            {
-                play_sub_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 3 ; i ++)
-        {
-            if(com_cruiser_pos[i] == 100)
-            {
-                play_cru_count++;
-                flag = 1;
-            }
-        }
-        for(int i = 0 ; i < 2 ; i ++)
-        {
-            if(com_destroyer_pos[i] == 100)
-            {
-                play_des_count++;
-                flag = 1;
-            }
-        }
-        if(flag == 1)
-        {
-            l100.setIcon(new ImageIcon("water_ship_hit.png"));
-            l100.setDisabledIcon(new ImageIcon("water_ship_hit.png"));
-            turn_flag = 0;
-            refresh_flag = 0;
-            play_hit_count++;
-        }
-        else
-        {
-            l100.setIcon(new ImageIcon("water_ship_miss.png"));
-            l100.setDisabledIcon(new ImageIcon("water_ship_miss.png"));
-            turn_flag = 1;
-            refresh_flag = 1;
-            miss();
-        }
-        l100.setEnabled(false);
-        who_is_playing();
-    }
-
-
-
-
 
     public void j1Call()
     {
